tech(api): refactor _createDirectKnowledgeElement for clarity
